requirements.txt
Django>=2.0.1
gunicorn==19.7.1
whitenoise==3.3.1
psycopg2==2.7.4
djangorestframework>=3.8.2
django-eventtools==0.9.11
django-widget-tweaks==1.4.1
django-filter==1.0.4





